+++ src/application/repository.ts
@@ -1,0 +1,11 @@
+export {
+  findRepositoryRoot,
+  GitSnapshotError,
+  openGitSnapshot,
+  openTrackedWorktree,
+  type GitSnapshotErrorCode,
+} from "../git.js";
+export {
+  openPackagedCase,
+  packagedCasePresentation,
+} from "../case.js";